fix(MongoConnectionHealthCheck): make the health check usable with both sync and async clients
