+++ cohort-mongo/build.gradle.kts
@@ -1,7 +1,8 @@
 dependencies {
    implementation(projects.cohortApi)
-   api(libs.mongodb.driver.sync)
+   compileOnly(libs.mongodb.driver.sync)
    testImplementation(libs.testcontainers.mongodb)
+   testImplementation(libs.mongodb.driver.sync)
 }
 
 apply("../publish.gradle.kts")
